+++ src/main/java/kg/ebooks/eBook/db/domain/model/others/Address.java
@@ -1,4 +1,4 @@
-package kg.ebooks.eBook.db.domain.model;
+package kg.ebooks.eBook.db.domain.model.others;
 
 import kg.ebooks.eBook.db.domain.model.enums.Country;
 import lombok.AllArgsConstructor;
