+++ src/main/java/kg/ebooks/eBook/db/domain/model/others/Basket.java
@@ -1,4 +1,4 @@
-package kg.ebooks.eBook.db.domain.model;
+package kg.ebooks.eBook.db.domain.model.others;
 
 import kg.ebooks.eBook.db.domain.model.books.Book;
 import lombok.AllArgsConstructor;
@@ -11,26 +11,25 @@ import java.util.List;
 
 import static javax.persistence.CascadeType.*;
 import static javax.persistence.CascadeType.PERSIST;
-import static javax.persistence.FetchType.EAGER;
+import static javax.persistence.FetchType.*;
 
 /**
  * created by Beksultan Mamatkadyr uulu
  * project : eBook
  * 9/1/22
- * Sunday 10:35
+ * Sunday 10:32
  */
 @Entity
-@Table(name = "selected_books")
+@Table(name = "baskets")
 @NoArgsConstructor @AllArgsConstructor
 @Getter @Setter
-public class SelectedBooks {
+public class Basket {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long selectedBooksId;
+    private Long basketId;
 
     @ManyToMany(fetch = EAGER, cascade = {DETACH, REFRESH, MERGE, PERSIST})
     private List<Book> books;
 
     private int quantityOfBooks;
-
 }
